requirements.txt
# Core
python==3.9
torch==2.1.0
torchvision==0.16.0

# Data & Computer Vision
opencv-python==4.8.1
pillow==10.1.0
facenet-pytorch==2.5.3
albumentations==1.3.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.4

# MLOps & API
mlflow==2.9.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Frontend
streamlit==1.29.0
plotly==5.18.0

# Utilities
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
ipywidgets==8.1.2

# Tests
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Other
pyyaml==6.0.1
python-dotenv==1.0.0